Fixed issue #20239: Reflected Cross-Site Scripting

+++ application/models/QuestionCreate.php
@@ -14,7 +14,7 @@ class QuestionCreate extends Question
         if (empty($oSurvey)) {
             throw new Exception('Found no survey with id ' . json_encode($iSurveyId));
         }
-        $gid = Yii::app()->request->getParam('gid', 0);
+        $gid = (int)Yii::app()->request->getParam('gid', 0);
         if ($gid == 0) {
             $gid = array_values($oSurvey->groups)[0]->gid;
         }
